Move NotificationRemoteMediator to separate file
